moved median metric to windown folder
