+++ ingestion/src/metadata/orm_profiler/metrics/registry.py
@@ -35,7 +35,7 @@ from metadata.orm_profiler.metrics.static.like_count import LikeCount
 from metadata.orm_profiler.metrics.static.max import Max
 from metadata.orm_profiler.metrics.static.max_length import MaxLength
 from metadata.orm_profiler.metrics.static.mean import Mean
-from metadata.orm_profiler.metrics.static.median import Median
+from metadata.orm_profiler.metrics.window.median import Median
 from metadata.orm_profiler.metrics.static.min import Min
 from metadata.orm_profiler.metrics.static.min_length import MinLength
 from metadata.orm_profiler.metrics.static.not_like_count import NotLikeCount
